+++ packages/common/src/accountTransactions.ts
@@ -9,6 +9,7 @@ import {
     serializeList,
     encodeWord8,
     serializeConfigureDelegationPayload,
+    serializeConfigureBakerPayload,
 } from './serializationHelpers';
 import {
     AccountTransactionType,
@@ -21,6 +22,7 @@ import {
     UpdateCredentialsPayload,
     RegisterDataPayload,
     ConfigureDelegationPayload,
+    ConfigureBakerPayload,
 } from './types';
 import { AccountAddress } from './types/accountAddress';
 import { DataBlob } from './types/DataBlob';
@@ -251,6 +253,26 @@ export class RegisterDataHandler
     }
 }
 
+export class ConfigureBakerHandler
+    implements AccountTransactionHandler<ConfigureBakerPayload>
+{
+    getBaseEnergyCost(payload: ConfigureBakerPayload): bigint {
+        if (payload.keys) {
+            return 4050n;
+        } else {
+            return 300n;
+        }
+    }
+
+    serialize(payload: ConfigureBakerPayload): Buffer {
+        return serializeConfigureBakerPayload(payload);
+    }
+
+    deserialize(): ConfigureBakerPayload {
+        throw new Error('deserialize not supported');
+    }
+}
+
 export class ConfigureDelegationHandler
     implements AccountTransactionHandler<ConfigureDelegationPayload>
 {
@@ -295,6 +317,9 @@ export function getAccountTransactionHandler(
     type: AccountTransactionType.ConfigureDelegation
 ): ConfigureDelegationHandler;
 export function getAccountTransactionHandler(
+    type: AccountTransactionType.ConfigureBaker
+): ConfigureBakerHandler;
+export function getAccountTransactionHandler(
     type: AccountTransactionType
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ): any {
@@ -315,6 +340,8 @@ export function getAccountTransactionHandler(
             return new RegisterDataHandler();
         case AccountTransactionType.ConfigureDelegation:
             return new ConfigureDelegationHandler();
+        case AccountTransactionType.ConfigureBaker:
+            return new ConfigureBakerHandler();
         default:
             throw new Error(
                 'The provided type does not have a handler: ' + type